requirements.txt
PyQt6>=6.6.0
SQLAlchemy>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
torch>=2.1.0
mpmath>=1.3.0

